Take 2, minimize the changes and preserve the Chunk Key.

+++ src/main/java/net/glowstone/io/StructureDataService.java
@@ -14,12 +14,12 @@ public interface StructureDataService {
      *
      * @return A map containing structures indexed by their chunk hash.
      */
-    Map<Long, GlowStructure> readStructuresData();
+    Map<Integer, GlowStructure> readStructuresData();
 
     /**
      * Write the structures data to storage.
      *
      * @param structures The structures to write to storage.
      */
-    void writeStructuresData(Map<Long, GlowStructure> structures);
+    void writeStructuresData(Map<Integer, GlowStructure> structures);
 }
